requirements.txt
Pillow==9.5.0
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Werkzeug==2.3.6